requirements.txt
streamlit>=1.25.0
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.3
numpy>=1.23.0
plotly>=5.15.0
pandas>=1.5.0 